+++ x/distribution/migrations/v2/store_test.go
@@ -58,7 +58,7 @@ func TestStoreMigration(t *testing.T) {
 		{
 			"DelegatorStartingInfo",
 			v1.GetDelegatorStartingInfoKey(valAddr, addr2),
-			types.GetDelegatorStartingInfoKey(valAddr, addr2),
+			append(append(types.DelegatorStartingInfoPrefix, address.MustLengthPrefix(valAddr.Bytes())...), address.MustLengthPrefix(addr2.Bytes())...),
 		},
 		{
 			"ValidatorHistoricalRewards",
@@ -73,7 +73,7 @@ func TestStoreMigration(t *testing.T) {
 		{
 			"ValidatorAccumulatedCommission",
 			v1.GetValidatorAccumulatedCommissionKey(valAddr),
-			types.GetValidatorAccumulatedCommissionKey(valAddr),
+			append(types.ValidatorAccumulatedCommissionPrefix, address.MustLengthPrefix(valAddr.Bytes())...),
 		},
 		{
 			"ValidatorSlashEvent",